mindmapp_app/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
deprecation==2.1.0
distro==1.9.0
dj-database-url==2.3.0
Django==4.2.16
django-cors-headers==4.5.0
django-environ==0.11.2
djangorestframework==3.15.2
exceptiongroup==1.2.2
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
google-api-core==2.21.0
google-auth==2.35.0
google-cloud-speech==2.27.0
googleapis-common-protos==1.65.0
gotrue==2.9.3
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.1
hyperframe==6.0.1
idna==3.10
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.2.1
numpy==2.0.2
openai==0.28.0
packaging==24.1
pillow==11.0.0
postgrest==0.17.2
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
realtime==2.0.6
regex==2024.9.11
requests==2.32.3
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
sentence-transformers==3.2.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
storage3==0.8.2
StrEnum==0.4.15
supabase==2.9.1
supafunc==0.6.2
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.20.1
torch==2.5.0
tqdm==4.66.5
transformers==4.45.2
typing_extensions==4.12.2
urllib3==2.2.3
websockets==13.1
yarl==1.16.0
gunicorn==20.0.4